Organize tests and remove scripts
